Break up models.py into models dir
