+++ helpdesk/models/FollowUpManager.py
@@ -1,0 +1,16 @@
+"""
+django-helpdesk - A Django powered ticket tracker for small enterprise.
+
+(c) Copyright 2008-2025 Jutda. All Rights Reserved. See LICENSE for details.
+
+models.py - Model (and hence database) definitions. This is the core of the
+            helpdesk structure.
+"""
+from django.db import models
+
+class FollowUpManager(models.Manager):
+    def private_followups(self):
+        return self.filter(public=False)
+
+    def public_followups(self):
+        return self.filter(public=True)
