+++ helpdesk/models/attachment_path.py
@@ -1,0 +1,11 @@
+"""
+django-helpdesk - A Django powered ticket tracker for small enterprise.
+
+(c) Copyright 2008-2025 Jutda. All Rights Reserved. See LICENSE for details.
+
+models.py - Model (and hence database) definitions. This is the core of the
+            helpdesk structure.
+"""
+def attachment_path(instance, filename):
+    """Just bridge"""
+    return instance.attachment_path(filename)
