+++ helpdesk/models/tickets_per_page_default.py
@@ -1,0 +1,12 @@
+"""
+django-helpdesk - A Django powered ticket tracker for small enterprise.
+
+(c) Copyright 2008-2025 Jutda. All Rights Reserved. See LICENSE for details.
+
+models.py - Model (and hence database) definitions. This is the core of the
+            helpdesk structure.
+"""
+from . import get_default_setting
+
+def tickets_per_page_default():
+    return get_default_setting("tickets_per_page")
